Ability to chose combating armies in the combat tool
Added an interface that allows the player to chose armies for both combatants in the combat tool in the dev tools section.

When a field is selected from one of the unit panels at the top of the screen, an editor window will pop up allowing the player to chose a unit to fill that field.
Player can:
   Change units
   Change number of units
   Change unit starting positions
   Delete units